majorProject2/requirements.txt
requests
python-dotenv
streamlit
opencv-contrib-python-headless
numpy
pillow